adjusted the data source implementations to i) the introduction of the parser interface and ii) the changes of the request processor interface
