+++ src/org/linkeddatafragments/datasource/hdt/HdtDataSource.java
@@ -2,25 +2,8 @@ package org.linkeddatafragments.datasource.hdt;
 
 import java.io.IOException;
 
-import org.linkeddatafragments.datasource.AbstractRequestProcessorForTriplePatterns;
 import org.linkeddatafragments.datasource.DataSource;
 import org.linkeddatafragments.datasource.IFragmentRequestProcessor;
-import org.linkeddatafragments.fragments.LinkedDataFragment;
-import org.linkeddatafragments.fragments.LinkedDataFragmentRequest;
-import org.linkeddatafragments.fragments.tpf.TriplePatternFragmentRequest;
-import org.rdfhdt.hdt.enums.TripleComponentRole;
-import org.rdfhdt.hdt.hdt.HDT;
-import org.rdfhdt.hdt.hdt.HDTManager;
-import org.rdfhdt.hdt.triples.IteratorTripleID;
-import org.rdfhdt.hdt.triples.TripleID;
-import org.rdfhdt.hdtjena.NodeDictionary;
-
-import com.hp.hpl.jena.graph.Triple;
-import com.hp.hpl.jena.rdf.model.Model;
-import com.hp.hpl.jena.rdf.model.ModelFactory;
-import com.hp.hpl.jena.rdf.model.Property;
-import com.hp.hpl.jena.rdf.model.RDFNode;
-import com.hp.hpl.jena.rdf.model.Resource;
 
 /**
  * An HDT data source of Basic Linked Data Fragments.
@@ -30,8 +13,7 @@ import com.hp.hpl.jena.rdf.model.Resource;
  */
 public class HdtDataSource extends DataSource {
 
-    private final HDT datasource;
-    private final NodeDictionary dictionary;
+    protected final HdtBasedRequestProcessorForTPFs requestProcessor;
 
     /**
      * Creates a new HdtDataSource.
@@ -43,98 +25,13 @@ public class HdtDataSource extends DataSource {
      */
     public HdtDataSource(String title, String description, String hdtFile) throws IOException {
         super(title, description);
-        datasource = HDTManager.mapIndexedHDT(hdtFile, null);
-        dictionary = new NodeDictionary(datasource.getDictionary());
+        requestProcessor = new HdtBasedRequestProcessorForTPFs( hdtFile );
     }
 
     @Override
-    public IFragmentRequestProcessor getRequestProcessor(
-            final LinkedDataFragmentRequest request )
+    public IFragmentRequestProcessor getRequestProcessor()
     {
-        if ( ! (request instanceof TriplePatternFragmentRequest) )
-            throw new IllegalArgumentException();
-
-        return new MyProcessor( (TriplePatternFragmentRequest) request );
+        return requestProcessor;
     }
 
-protected class MyProcessor extends AbstractRequestProcessorForTriplePatterns
-{
-    public MyProcessor( final TriplePatternFragmentRequest request ) {
-        super( request );
-    }
-
-    @Override
-    protected LinkedDataFragment createFragment( final Resource subject,
-                                                 final Property predicate,
-                                                 final RDFNode object,
-                                                 final long offset,
-                                                 final long limit )
-    {
-        // look up the result from the HDT datasource)
-        int subjectId = subject == null ? 0 : dictionary.getIntID(subject.asNode(), TripleComponentRole.SUBJECT);
-        int predicateId = predicate == null ? 0 : dictionary.getIntID(predicate.asNode(), TripleComponentRole.PREDICATE);
-        int objectId = object == null ? 0 : dictionary.getIntID(object.asNode(), TripleComponentRole.OBJECT);
-        
-        if (subjectId < 0 || predicateId < 0 || objectId < 0) {
-            return createEmptyTriplePatternFragment();
-        }
-        
-        final Model triples = ModelFactory.createDefaultModel();
-        IteratorTripleID matches = datasource.getTriples().search(new TripleID(subjectId, predicateId, objectId));
-        boolean hasMatches = matches.hasNext();
-		
-        if (hasMatches) {
-            // try to jump directly to the offset
-            boolean atOffset;
-            if (matches.canGoTo()) {
-                try {
-                    matches.goTo(offset);
-                    atOffset = true;
-                } // if the offset is outside the bounds, this page has no matches
-                catch (IndexOutOfBoundsException exception) {
-                    atOffset = false;
-                }
-            } // if not possible, advance to the offset iteratively
-            else {
-                matches.goToStart();
-                for (int i = 0; !(atOffset = i == offset) && matches.hasNext(); i++) {
-                    matches.next();
-                }
-            }
-            // try to add `limit` triples to the result model
-            if (atOffset) {
-                for (int i = 0; i < limit && matches.hasNext(); i++) {
-                    triples.add(triples.asStatement(toTriple(matches.next())));
-                }
-            }
-        }
-
-        // estimates can be wrong; ensure 0 is returned if there are no results, 
-        // and always more than actual results
-        final long estimatedTotal = triples.size() > 0 
-                ? Math.max(offset + triples.size() + 1, matches.estimatedNumResults())
-                : hasMatches 
-                    ? Math.max(matches.estimatedNumResults(), 1) 
-                    : 0;
-
-        // create the fragment
-        final boolean isLastPage = ( estimatedTotal < offset + limit );
-        return createTriplePatternFragment( triples, estimatedTotal, isLastPage );
-    }
-
-} // end of MyProcessor
-
-    /**
-     * Converts the HDT triple to a Jena Triple.
-     *
-     * @param tripleId the HDT triple
-     * @return the Jena triple
-     */
-    private Triple toTriple(TripleID tripleId) {
-        return new Triple(
-            dictionary.getNode(tripleId.getSubject(), TripleComponentRole.SUBJECT),
-            dictionary.getNode(tripleId.getPredicate(), TripleComponentRole.PREDICATE),
-            dictionary.getNode(tripleId.getObject(), TripleComponentRole.OBJECT)
-        );
-    }
 }
